requirements.txt
mss
Pillow
pystray
keyboard
# tkinter (standard library)
